Fix crashes reported on Crashytics
Crashlytics issues:
- 5691d893f5d3a7f76bf932e8
- 5692a9cef5d3a7f76bfc66a0
- 56937e16f5d3a7f76bff7583
- 56951250f5d3a7f76b06d82b

+++ app/src/main/java/org/exoplatform/utils/CompatibleFileOpen.java
@@ -234,7 +234,7 @@ public class CompatibleFileOpen {
         return RESULT_OK;
       } catch (IOException e) {
         Log.d(getClass().getSimpleName(), e.getMessage(), Log.getStackTraceString(e));
-        if (file != null) {
+        if (file != null && file.exists()) {
           file.delete();
         }
         return RESULT_ERROR;
@@ -248,10 +248,10 @@ public class CompatibleFileOpen {
      */
     @Override
     protected void onCancelled() {
-      if (file.exists()) {
+      if (file != null && file.exists()) {
         file.delete();
       }
-      if (mProgressDialog.isAttachedToWindow())
+      if (mProgressDialog != null && mProgressDialog.isAttachedToWindow())
         mProgressDialog.dismiss();
       super.onCancelled();
     }
